Expose empty Query type
In the current version of GraphQL, you can’t have an empty type even if
you intend to extend it later. So we need to make sure the Query type
has at least one field — in this case we can add a fake _empty field.
Hopefully in future versions it will be possible to have an empty type
to be extended later.

[Link](https://blog.apollographql.com/modularizing-your-graphql-schema-code-d7f71d5ed5f2#14ac)

Co-authored-by: Guilherme Eiras <guialbertoeiras@hotmail.com>